feat : Type 클래스들을 type 디렉토리로 이동
