+++ BE/src/main/java/bibi/demo/domain/type/Allergen.java
@@ -1,16 +1,14 @@
-package bibi.demo.domain;
+package bibi.demo.domain.type;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import javax.persistence.*;
 
-@Entity
 @Getter
-public class ToppingType {
+@Entity
+@Table(name = "allergen")
+public class Allergen {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -20,10 +18,10 @@ public class ToppingType {
     @JsonProperty("name_en")
     private String nameEN;
 
-    public ToppingType() {
+    public Allergen() {
     }
 
-    public ToppingType(Long id, String nameKR, String nameEN) {
+    public Allergen(Long id, String nameKR, String nameEN) {
         this.id = id;
         this.nameKR = nameKR;
         this.nameEN = nameEN;
